表示項目を修正する
